Return null when an EOFException occurs...
instead of an IOException, which causes a platform based server to
return a 500 error.

The EOFException occurs when the input is blank (e.g. a blank string).
It seems reasonable that when the server encounters a blank string, the
value should be null instead of an IOException, however I'm not exactly
sure if this exception occurs elsewhere, and i don't know how to test it
completely.