Add additional columns to the HTTP server logs
This change adds the following columns to the HTTP server logs:

- stream id: stream identifier for HTTP/2 requests, -1 otherwise.
- protocol version: version of the HTTP protocol specified by the
request.
- request begin-to-dispatch time [ms]: time from when the request is
received to when the request is dispatched to the application.
- request begin-to-end time [ms]: time from when the request is received
to when the request has been fully consumed/parsed.
- first-to-last-content-time [ms]: time from when the first chunk of the
response content has been written to the network to when the last
chunk of the response content has been written.
- response content interarrival time stats: the summary statistics
for the interarrival time of the response content writes. For example,
if three response content chunks have been written by the HTTP server
at t=0, t=3, and t=20, then this field will contain the summary
statistics for the values 3 and 17.